Refactor: login
